GlobalExceptionHandler and Unit Testing

+++ Company/src/main/java/com/epam/service/impl/CompanyServiceImpl.java
@@ -1,8 +1,12 @@
 package com.epam.service.impl;
 
+import java.lang.reflect.Parameter;
 import java.util.List;
 
 import com.epam.dto.CompanyDto;
+import com.epam.exception.DuplicateDataException;
+import com.epam.exception.NotFoundException;
+import com.epam.exception.ParameterNotCorrectException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,19 +30,31 @@ public class CompanyServiceImpl implements CompanyService {
 	}
 
 	@Override
-	public Company getCompanyByName(String name) {
+	public Company getCompanyByName(String name) throws ParameterNotCorrectException, NotFoundException {
+		if(name.isBlank())
+			throw new ParameterNotCorrectException("Name cannot be blank");
+		if(!companyRepo.existsByName(name))
+			throw new NotFoundException("Company with name " + name + " is not found in database");
+
 		return companyRepo.findByName(name);
 	}
 
 	@Override
-	public boolean addCompany(CompanyDto companyDto) {
+	public boolean addCompany(CompanyDto companyDto) throws DuplicateDataException {
+		String name = companyDto.getName();
+		if(companyRepo.existsByName(name))
+			throw new DuplicateDataException("Company with Name " + name + " is already present in the database");
+
 		Company company = modelMapper.map(companyDto,Company.class);
 		companyRepo.save(company);
 		return true;
 	}
 
 	@Override
-	public boolean removeCompany(String name) {
+	public boolean removeCompany(String name) throws NotFoundException {
+		if(!companyRepo.existsByName(name))
+			throw new NotFoundException("Company with name " + name + " is not found in database");
+
 		Company company = companyRepo.findByName(name);
 		companyRepo.delete(company);
 		return true;
